@@ -1,6 +1,6 @@
 {
   "name": "koa",
-  "version": "2.5.3",
+  "version": "2.7.0",
   "description": "Koa web app framework",
   "main": "lib/application.js",
   "scripts": {
@@ -48,7 +48,7 @@
     "vary": "^1.1.2"
   },
   "devDependencies": {
-    "eslint": "^3.17.1",
+    "eslint": "^6.0.1",
     "eslint-config-koa": "^2.0.0",
     "eslint-config-standard": "^7.0.1",
     "eslint-plugin-promise": "^3.5.0",
