@@ -87,7 +87,7 @@ describe('app.respond', () => {
   });
 
   describe('when this.type === null', () => {
-    it('should not send Content-Type header', async () => {
+    it('should not send Content-Type header', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -106,7 +106,7 @@ describe('app.respond', () => {
   });
 
   describe('when HEAD is used', () => {
-    it('should not respond with the body', async () => {
+    it('should not respond with the body', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -124,7 +124,7 @@ describe('app.respond', () => {
       assert(!res.text);
     });
 
-    it('should keep json headers', async () => {
+    it('should keep json headers', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -142,7 +142,7 @@ describe('app.respond', () => {
       assert(!res.text);
     });
 
-    it('should keep string headers', async () => {
+    it('should keep string headers', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -160,7 +160,7 @@ describe('app.respond', () => {
       assert(!res.text);
     });
 
-    it('should keep buffer headers', async () => {
+    it('should keep buffer headers', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -301,7 +301,7 @@ describe('app.respond', () => {
     });
 
     describe('with status=204', () => {
-      it('should respond without a body', async () => {
+      it('should respond without a body', async() => {
         const app = new Koa();
 
         app.use(ctx => {
@@ -320,7 +320,7 @@ describe('app.respond', () => {
     });
 
     describe('with status=205', () => {
-      it('should respond without a body', async () => {
+      it('should respond without a body', async() => {
         const app = new Koa();
 
         app.use(ctx => {
@@ -339,7 +339,7 @@ describe('app.respond', () => {
     });
 
     describe('with status=304', () => {
-      it('should respond without a body', async () => {
+      it('should respond without a body', async() => {
         const app = new Koa();
 
         app.use(ctx => {
@@ -358,7 +358,7 @@ describe('app.respond', () => {
     });
 
     describe('with custom status=700', () => {
-      it('should respond with the associated status message', async () => {
+      it('should respond with the associated status message', async() => {
         const app = new Koa();
         statuses['700'] = 'custom status';
 
@@ -378,7 +378,7 @@ describe('app.respond', () => {
     });
 
     describe('with custom statusMessage=ok', () => {
-      it('should respond with the custom status message', async () => {
+      it('should respond with the custom status message', async() => {
         const app = new Koa();
 
         app.use(ctx => {
@@ -416,7 +416,7 @@ describe('app.respond', () => {
   });
 
   describe('when .body is a null', () => {
-    it('should respond 204 by default', async () => {
+    it('should respond 204 by default', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -433,7 +433,7 @@ describe('app.respond', () => {
       assert.equal(res.headers.hasOwnProperty('content-type'), false);
     });
 
-    it('should respond 204 with status=200', async () => {
+    it('should respond 204 with status=200', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -451,7 +451,7 @@ describe('app.respond', () => {
       assert.equal(res.headers.hasOwnProperty('content-type'), false);
     });
 
-    it('should respond 205 with status=205', async () => {
+    it('should respond 205 with status=205', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -469,7 +469,7 @@ describe('app.respond', () => {
       assert.equal(res.headers.hasOwnProperty('content-type'), false);
     });
 
-    it('should respond 304 with status=304', async () => {
+    it('should respond 304 with status=304', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -522,7 +522,7 @@ describe('app.respond', () => {
   });
 
   describe('when .body is a Stream', () => {
-    it('should respond', async () => {
+    it('should respond', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -541,7 +541,7 @@ describe('app.respond', () => {
       assert.deepEqual(res.body, pkg);
     });
 
-    it('should strip content-length when overwriting', async () => {
+    it('should strip content-length when overwriting', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -561,7 +561,7 @@ describe('app.respond', () => {
       assert.deepEqual(res.body, pkg);
     });
 
-    it('should keep content-length if not overwritten', async () => {
+    it('should keep content-length if not overwritten', async() => {
       const app = new Koa();
 
       app.use(ctx => {
@@ -582,7 +582,7 @@ describe('app.respond', () => {
     });
 
     it('should keep content-length if overwritten with the same stream',
-      async () => {
+      async() => {
         const app = new Koa();
 
         app.use(ctx => {
@@ -777,7 +777,7 @@ describe('app.respond', () => {
         .expect('hello');
     });
 
-    it('should 204', async () => {
+    it('should 204', async() => {
       const app = new Koa();
 
       app.use(ctx => {
